requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
dateparser==1.1.1
express==0.1.5
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
node==1.0
npm==0.1.1
numpy==1.22.4
odict==1.9.0
optional-django==0.1.0
pandas==1.4.2
plumber==1.7
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.28.0
six==1.16.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.4
zope.proxy==4.5.0
